Load storymap content with chunk
This fixes the loading sequence of any JS assets added from inside the content area.

When using the regular [[*content]] placeholder, the loadAssets snippets inside the content area are parsed _after_ the [[$script]] tag at the bottom of the template. This leads to the site.js file being loaded before some of the JS dependencies that it initializes (such as tab and embed).