Manage Places from storymap Chapter interface
WIP. The Add Place content button is cloned from the MIGX TV grid, but it doesn't clone the attached events yet.

That should be the case with a Jquery deep clone, but the ice is a bit slippery here because many things are loaded async or via AJAX in the manager.